requirements.txt
Flask==3.0.3
flask_cors==3.0.10
datetime==4.3.0
python-dotenv==0.20.0
cohere==4.26.1